pyproject.toml
[project]
name = "fastapi-ca"
version = "0.1.0"
description = ""
authors = [
    {name = "dextto", email = "dexter.haan@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.25.0",
    "py-ulid>=1.0.3",
    "passlib[bcrypt]>=1.7.4",
    "sqlalchemy>=2.0.25",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.1",
    "dependency-injector>=4.41.0",
    "pydantic[email]>=2.5.3",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.1.0",
    "celery>=5.3.6",
    "requests>=2.31.0",
    "pytest>=8.0.0",
    "freezegun>=1.4.0",
    "pytest-mock>=3.12.0"
]

[tool.setuptools.packages.find]
include = ["*"]
exclude = ["migrations*", "example*"]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
